Move database files to database folder
