+++ src/core/core.module.ts
@@ -1,7 +1,7 @@
 import { Global, Module } from '@nestjs/common';
 import { AwsS3Factory } from './aws-s3.factory';
 import { ConfigModule } from './config/config.module';
-import { CypherFactory } from './cypher.factory';
+import { CypherFactory } from './database/cypher.factory';
 import { DatabaseService } from './database.service';
 import { PropertyUpdaterService } from './database/property-updater.service';
 import { IndexerModule } from './database/indexer/indexer.module';
